populate common time.h, remove rp6502 clock.h
