+++ mos-platform/rp6502/clock_getres.c
@@ -1,5 +1,5 @@
-#include <clock.h>
 #include <rp6502.h>
+#include <time.h>
 
 int clock_getres(clockid_t clock_id, struct timespec *res) {
   RIA.a = clock_id;
